Fixed scroll bug on logout

+++ frontend/components/message/message_list.jsx
@@ -14,6 +14,9 @@ class MessageList extends React.Component {
 
   render(){
     let that = this;
+    if (!this.props.currentUser){
+      return (<div></div>);
+    }
     // debugger
     let messageKeys = Object.keys(this.props.messages);
     let content = messageKeys.map(key=>{
@@ -23,13 +26,13 @@ class MessageList extends React.Component {
       let button=(<span></span>);
       let message = this.props.messages[key];
       let date = new Date(message.date);
-      if (this.props.currentUser.id === message.user_id){
+      if (this.props.currentUser && this.props.currentUser.id === message.user_id){
         button = (
           <button className="msg-delete-button" onClick={this.handleDelete(this.props.messages[key].id)}>Delete</button>
         );
       }
       return (
-        <li className="msg-list-item group" key={`${key} ${message.date}`}><span className="message-info group"><span className="msg-username">{this.props.users[message.user_id] ? this.props.users[message.user_id].username : currentUser.username}</span><span className='message-date'>{date.toTimeString()}</span>{button}</span>{message.body}</li>
+        <li className="msg-list-item group" key={`${key} ${message.date}`}><span className="message-info group"><span className="msg-username">{this.props.users[message.user_id] ? this.props.users[message.user_id].username : this.props.currentUser.username}</span><span className='message-date'>{date.toTimeString()}</span>{button}</span>{message.body}</li>
       );
     });
     return (
